Get rid of some more insidious longs.
Especially constructs like: long foo = x.props[].get_int()
can get you by surprise on 32 bit arches -- even worse now that most
devs are on amd64.

Item flags I typedeffed as iflags_t, as it's likely we'll have to extend
it soon, yet defining it as uint64_t now would be misleading since there
are so many places it's used in a 32 bit manner.

+++ crawl-ref/source/tileview.cc
@@ -682,7 +682,7 @@ void tile_place_monster(const coord_def &gc, const monster* mon)
         return;
     else if (pref == TAGPREF_TUTORIAL)
     {
-        const long kills = you.kills->num_kills(mon);
+        const int kills = you.kills->num_kills(mon);
         const int limit  = 0;
 
         if (!mon->is_named() && kills > limit)
